// frontend/src/types/contacts.types.ts
export interface Contact {
  id?: string;
  email?: string;
  notes?: string;
  github?: string;
  address?: string;
  twitter?: string;
  jobTitle?: string;
  lastName?: string;
  createdBy?: string;
  birthday?: string;
  facebook?: string;
  linkedin?: string;
  firstName?: string;
  department?: string;
  phoneNumber?: string;
  communicationChannel?: string;
}